smart_door_lock_backend/requirements.txt
Flask==2.2.5
mysql-connector-python
requests==2.31.0
